% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_caplter.R
\docType{data}
\name{capAddress}
\alias{capAddress}
\title{default CAP LTER address}
\format{
An object of class \code{list} of length 8.
}
\usage{
capAddress
}
\description{
default CAP LTER address
}
\keyword{datasets}
