% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_caplter.R
\docType{data}
\name{capContact}
\alias{capContact}
\title{default CAP LTER contact as EML::contact}
\format{
An object of class \code{list} of length 11.
}
\usage{
capContact
}
\description{
default CAP LTER contact as EML::contact
}
\keyword{datasets}
