% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_caplter.R
\docType{data}
\name{capPublisher}
\alias{capPublisher}
\title{default CAP LTER publisher as EML::publisher}
\format{
An object of class \code{list} of length 12.
}
\usage{
capPublisher
}
\description{
default CAP LTER publisher as EML::publisher
}
\keyword{datasets}
