% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_caplter.R
\name{capRights}
\alias{capRights}
\title{capRights: default intellectual rights statement passed to EML-generating workflow}
\usage{
capRights
}
\description{
capRights: default intellectual rights statement passed to EML-generating workflow
}
