% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_caplter_project.R
\name{configure_caplter_project}
\alias{configure_caplter_project}
\title{Configure CAP LTER project details for a EML project entity}
\usage{
configure_caplter_project()
}
\value{
establishes in the R environment a EML::project entity specific to
the CAP LTER.
}
\description{
configure_caplter_project establishes an EML::project with the
elements: (1) personnel (CAP PI and Co-PIs), (2) project abstract (project
overview from CAP IV proposal), and (3) funding (grant numbers for CAP I,
II, III, and IV).
}
\details{
configure_caplter_project establishes an EML::project with the
elements: (1) personnel (CAP PI and Co-PIs), (2) project abstract (project
overview from CAP IV proposal), and (3) funding (grant numbers for CAP I,
II, III, and IV).
}
