% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_som_project.R
\name{configure_som_project}
\alias{configure_som_project}
\title{Configure SOM project details for a EML project entity}
\usage{
configure_som_project()
}
\value{
establishes in the R environment upon package load a EML::project
entity specific to the LTER SOM Working Group.
}
\description{
configure_som_project establishes an EML::project with the
elements: (1) personnel, (2) project abstract, and (3) funding (grant
numbers)
}
\details{
configure_som_project establishes an EML::project with the
elements: (1) personnel, (2) project abstract, and (3) funding (grant
numbers).
}
