% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_urex_project.R
\name{configure_urex_project}
\alias{configure_urex_project}
\title{Configure UREx project details for a EML project entity}
\usage{
configure_urex_project()
}
\value{
Establishes in the R environment a EML::project entity specific to
the UREx project.#'
}
\description{
configure_urex_project establishes an EML::project with the
elements: (1) personnel, (2) project abstract, and (3) funding
}
\details{
configure_urex_project establishes an EML::project with the
elements: (1) personnel, (2) project abstract, and (3) funding
}
