% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_caplter.R
\name{create_distribution}
\alias{create_distribution}
\title{create_distribution}
\usage{
create_distribution(packageIdent)
}
\arguments{
\item{packageIdent}{The data package number}
}
\description{
create_distribution
}
