% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_eml.R
\name{create_eml}
\alias{create_eml}
\title{create a EML entity of type EML}
\usage{
create_eml()
}
\value{
EML entity of type EML is returned.
}
\description{
\code{create_eml} generates a EML entity of type eml from
objects and metadata compiled from capeml objects in the working directory
}
\details{
A eml entity is created from objects defined in the user's R
environment and configuration files.
}
\note{
\code{create_eml} will look for most inputs used to construct a eml
entity, such as access and dataset, in the working environment; these
parameters are not passed directly to the function and must exist in the
working environment.

Some parameters, such access, are loaded in the backgroud when the
package is loaded and not called directly by the user.
}
