% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_geographic_coverage.R
\name{create_geographic_coverage}
\alias{create_geographic_coverage}
\title{Generate geographic coverage element(s) for one or more locations
from a simple features object}
\usage{
create_geographic_coverage(sf_object, description)
}
\arguments{
\item{sf_object}{ignored}

\item{description}{ignored}
}
\value{
Stops with a deprecation error.
}
\description{
\strong{Deprecated.} This function has moved to the
\code{capemlVector} package. Please install \code{capemlVector} and call
\code{capemlVector::create_geographic_coverage()} instead.
}
\note{
This function was relocated to the \code{capemlVector} package to
allow the \code{capeml} package to remove its dependency on \code{sf}.
Install the replacement with:
\code{pak::pak("CAPLTER/capemlVector")}
}
