% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_keyword.R
\name{create_keyword}
\alias{create_keyword}
\title{create_keyword}
\usage{
create_keyword(datasetKeyword, theAttribute)
}
\arguments{
\item{datasetKeyword}{A keyword or subject.}

\item{theAttribute}{A keyword attribute.}
}
\value{
EML keyword object is returned.
}
\description{
create_keyword generates a EML entity of type keyword with the
  option of adding an attribute to the keyword XML element. This is primarily
  inteneded as a helper function for the create_keywordSet function but can
  be run independently.
}
\details{
A EML entity of type keyword is created from a string (keyword
  element) and optional string as the keyword attribute.
}
\examples{
\dontrun{

arizona <- create_keyword('Arizona', 'place')
avifauna <- create_keyword('avifauna', 'theme')
phoenix <- create_keyword('phoenix')

}

}
