% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_KML.R
\name{create_KML}
\alias{create_KML}
\title{create_KML}
\usage{
create_KML(kmlFile, description)
}
\arguments{
\item{kmlFile}{The quoted name and path of the kml file.}

\item{description}{A description of the data.}
}
\value{
EML entity of type otherEntity with a kml dataType is returned.
  Additionally, the spatial data file is renamed with the project id + base
  file name + md5sum + file extension (kml).
}
\description{
create_KML generates a EML entity of type otherEntity with a kml
  dataType
}
\details{
a otherEntity entity is created from a single kml file (e.g.,
  desert_fertilization_sampling_sites.kml). The resulting entity is renamed
  with the project id + base file name + md5sum + file extension (kml).
}
\note{
create_KML will look for a project id in the working environment; this
  parameter is not passed to the function and it must exist.

create_KML currently does not accept an argument for url to a data
  file; the package defaults to a CAP LTER specific URL, and this must be
  changed manually in the resulting xml

the source kml file can be located anywhere on a local computer but the
  renamed file with project id and hash will be written to the current
  working directory
}
\examples{
\dontrun{

 # source file can reside in or out of the working directory (but the output
 file will be written to the working directory)

 desert_fertilization_sites <- create_KML(
   kmlFile = "~/Desktop/desert_fertilization_sampling_sites.kml",
   description = "approximate location of desert fertiliztion long-term study sites")

 desert_fertilization_sites <- create_KML(
   kmlFile = "desert_fertilization_sampling_sites.kml",
   description = "approximate location of desert fertiliztion long-term study sites")

}

}
