% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_spatialVector.R
\name{create_spatialVector}
\alias{create_spatialVector}
\title{create_spatialVector}
\usage{
create_spatialVector(svname, description,
  baseURL = "https://data.gios.asu.edu/datasets/cap/")
}
\arguments{
\item{svname}{The unquoted name of the spatial data object in the R
environment.}

\item{description}{A quoted description of the spatial object that will
populate the entityDescription element of the EML document.}

\item{baseURL}{(optional) The base path of the web-accessible location of the
data file; the name of the resulting file will be passed to the base path
to generate a web-resolvable file path.}
}
\value{
EML spatialVector object is returned. Additionally, the spatial data
  entity is written to file as type kml, and renamed with the project id +
  base file name + file extension (kml in this case).
}
\description{
create_spatialVector generates a EML entity of type
  spatialVector
}
\details{
create_spatialVector creates a EML spatialVector object from a
  spatial data object (shapefile, kml) that is read into the R environment.
  The function reads the attributes and classes contained within a supporting
  csv file generated from the write_attributes function -
  create_spatialVector will look for a file in the working directory with a
  name of type spatialEntityName_attrs.csv. Because all attributes written to
  a kml (the output of create_spatialVector  are of type character, factors
  are not considered in this workflow). Note that this functionality is
  predicated on the existence of a file containing metadata about the
  attributes, that that file is in the working directory, and that the file
  matches the name of the spatial data entity precisely. In addition to
  generating a EML entity of type spatialVector, create_spatialVector writes
  the spatial entity in the R environment to file as type kml, renames the
  file with project id + base file name + file extension (kml in this case).
}
\note{
create_spatialVector will look for a project id in the working environment;
  this parameter is not passed to the function and it must exist.

create_dataTable currently accepts an argument for a base url path to
  which the new file name will be appended so as to be a web-resolvable file;
  the package defaults to a URL specific to the CAP LTER.

Because file name metadata is incorporated into the kml generated from
  the sf::st_write function, including the md5 hash of the object in the file
  name sensu other products generated by this package is not feasible for
  spatialVector objects.

All vector objects are transformed to epsg 4326 (WGS 1984)
}
\examples{
\dontrun{

 spatialVectorEntity_SV <- create_spatialVector(
   svname = data object in R env.,
   description = "data from the field study")

}

}
