% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emlCoordSystems-data.R
\docType{data}
\name{emlCoordSystems}
\alias{emlCoordSystems}
\title{EML-compliant coordinate system names}
\format{
data frame
}
\source{
\url{https://knb.ecoinformatics.org/#external//emlparser/docs/eml-2.1.1/./eml-spatialReference.html}
}
\usage{
emlCoordSystems
}
\description{
EML has a defined list of coordinate systems that are permissible to generate
valid EML. emlCoordSystems is a list (as a data frame) of those string
values. The get_emlProjection() function attempts to match the coordinate
system metadata inherent in a spatial data entity with the corresponding
EML-compliant name listed in emlCoordSystems.
}
\keyword{datasets}
