% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annotations.R
\name{get_annotations}
\alias{get_annotations}
\title{translate annotations metadata to eml}
\usage{
get_annotations()
}
\value{
`EML::eml$annotations`
}
\description{
\code{get_annotations} harvests metadata pertaining to
annotations in a project `annotations.yaml` file and translates those
metadata to a list of type `EML::eml$annotation` for inclusion in a <eml>
element.
}
