% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_next_version.R
\name{get_next_version}
\alias{get_next_version}
\title{get the next version of a data package}
\usage{
get_next_version(
  provided_scope = "knb-lter-cap",
  provided_identifier,
  display_message = FALSE
)
}
\arguments{
\item{provided_scope}{(character) scope of data package (defaults to CAP LTER: knb-lter-cap)}

\item{provided_identifier}{(integer) identifier of data package}

\item{display_message}{(boolean) indicates whether to display a message if an existing dataset is
not identifed in the repository (thus returning a version number 1)}
}
\value{
integer reflecting the next version of a data package
}
\description{
\code{get_next_version} will access the EDI API to identify and
return the next version of a data package. If the data package does not
exist (i.e., we are constructing a new package), \code{get_next_version}
will return `1` and display a corresponding note.
}
\details{
`get_next_version()` authenticates automatically with EDI using
the `EDI_USER` and `EDI_PASS` environment variables when needed.

\code{get_next_version} is a helper function designed to aid
construction of a package identifier (e.g., the version `5` of
`knb-lter-cap.624.5`). As such, the most common use case is that the
function will be called internally from \code{capeml::createDataset} but the
function can be called directly.
}
\examples{
\dontrun{

get_next_version(
  provided_scope      = "knb-lter-cap",
  provided_identifier = 624
)

}
#'

}
