% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_package_evaluation.R
\name{get_package_evaluation}
\alias{get_package_evaluation}
\title{evaluate data package with EDI API}
\usage{
get_package_evaluation(identifier = NULL, scope = NULL, full_report = FALSE)
}
\arguments{
\item{identifier}{(integer) Unquoted package identifier integer}

\item{scope}{(character) Quoted package scope (e.g., "knb-lter-cap", "edi")}

\item{full_report}{(logical) Logical indicating whether to return a summary (default) or
detailed evaluation report.}
}
\value{
summary or detailed evaluation report
}
\description{
\code{get_package_evaluation} will use the generated eml.xml
and uploaded data to check the congruence of a data package.
}
\note{
This is a full evaluation of the data package, including checking
supplied metadata against provided data, as opposed to, for example, just
checking that the schema is valid. As such, the data files must be available
to EDI for proper evaluation.

\code{get_package_evaluation} will look for a config.yaml file to
estalish the package scope and identifier. Alternatively, these parameters
can be provided by the user but, in that case, both parameters must be
provided.

`get_package_evaluation()` automatically authenticates with EDI using
the `EDI_USER` and `EDI_PASS` environment variables when an EDI session is
not already active.

If warns or errors are returned, we can view the full report to
examine the details of those problems. For example, write the report to file
with \code{xml2::write_xml(report, "/tmp/report.xml")}, and search for e.g.,
`<status>warn`.
}
