% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_taxa_map.R
\name{get_taxa_db}
\alias{get_taxa_db}
\title{Connect to a local taxadb database}
\usage{
get_taxa_db()
}
\value{
A DBI database connection.
}
\description{
Helper function to connect to a local taxadb database and ensure the ITIS
schema is present.
}
\keyword{internal}
