% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_caplter.R
\docType{data}
\name{giosContact}
\alias{giosContact}
\title{default GIOS contact as EML::contact}
\format{An object of class \code{list} of length 11.}
\usage{
giosContact
}
\description{
default GIOS contact as EML::contact
}
\keyword{datasets}
