% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_caplter.R
\name{giosLanguage}
\alias{giosLanguage}
\title{giosLanguage: default language (english) passed to EML-generating workflow}
\usage{
giosLanguage
}
\description{
giosLanguage: default language (english) passed to EML-generating workflow
}
