% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_resolvable_taxa.R
\name{identify_resolvable_taxa}
\alias{identify_resolvable_taxa}
\title{Identify taxa within a list that can be matched to ITIS}
\usage{
identify_resolvable_taxa(taxa_list, nameType, interactive = FALSE)
}
\arguments{
\item{taxa_list}{A list of taxa to validate}

\item{nameType}{Indiate whether the taxonomic nomenclature refers to common
or scientific names}

\item{interactive}{Indiate whether the resolving taxa should be interactive.
If TRUE and more than one TSN is found for the species, the user is asked
for input. If FALSE (default) NA is returned for multiple matches.}
}
\value{
List of provided taxa and associated ITIS identifier if a match was
  found (NA if otherwise).
}
\description{
Provided with a list of taxa names, identify_resolvable_taxa
  generates a list of those taxa and the associated ITIS identifier if a
  match was identified. identify_resolvable_taxa is intended primarily as a
  helpful function for set_taxonomicCoverage but can be used independently.
}
\details{
A list of provided taxa and associated ITIS identifier if a match
  could be found.
}
\examples{
\dontrun{

allTaxa <- identify_resolvable_taxa(taxa_list = c('Ephemeroptera', 'Plecoptera'),
                                    nameType = "scientific",
                                    interactive = TRUE)

allTaxa <- identify_resolvable_taxa(taxa_list = 'Ephemeroptera')
                                    nameType = "scientific",

datasetTaxa <- core_birds \%>\%
  distinct(taxaColumn) \%>\%
  filter(!is.na(taxaColumn)) \%>\%
  pull(taxaColumn)

allTaxa <- identify_resolvable_taxa(taxa_list = datasetTaxa,
                                    nameType = "common")

}

}
