% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_caplter.R
\docType{data}
\name{lterAccess}
\alias{lterAccess}
\title{default CAP LTER / GIOS access as EML::access}
\format{
An object of class \code{list} of length 8.
}
\usage{
lterAccess
}
\description{
default CAP LTER / GIOS access as EML::access
}
\keyword{datasets}
