% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_attributes.R
\name{read_attributes}
\alias{read_attributes}
\title{Read information from attributes and factors metadata files, and
document missing value codes}
\usage{
read_attributes(entity_name, missing_value_code = NULL, return_type = "eml")
}
\arguments{
\item{entity_name}{(character) The quoted name of the data entity.}

\item{missing_value_code}{(optional)
(character) read_attributes will automatically document the presence of NA
and NaN entries as missing values in the EML output. The user has the
ability to identify an additional indicator of missing values (e.g.,
"-9999", "missing") if present.}

\item{return_type}{(character) Quoted designator indicating the value returned as either an EML
attributes entity (return_type = "eml", the default) or a dataframe of
entity attributes and column classes (return_type = "attributes") read from
the attributes file, the latter primarily as a helper feature for updating
an existing attributes file.}
}
\value{
A entity of type EML attributes
}
\description{
The read_attributes function reads an entity's attribute
details from a "entity name"_attrs.yaml or "entity name"_attrs.csv file in
the working directory - the details of which are incorporated into the EML
metadata for that entity. If present in the working directory, factor
metadata in a "entity name"_factors.yaml or "entity name"_factors.csv are
read and incorporated into attribute metadata. NA, NaN, and user-provided
missing value codes are documented for each variable if they exist.
}
\details{
The read_attributes function reads an entity's attribute
details from a "entity name"_attrs.yaml or "entity name"_attrs.csv file in
the working directory - the details of which are incorporated into the EML
metadata for that entity. If present in the working directory, factor
metadata in a "entity name"_factors.yaml or "entity name"_factors.csv are
read and incorporated into attribute metadata. NA, NaN, and user-provided
missing value codes are documented for each variable if they exist.
}
\note{
Recent versions of the capeml package generate attribute and factor
metadata files in yaml format; the read_attributes function will look also
for attributes files in csv format to accommodate older projects.

The read_attributes function is intended primarily as a helper to
other functions in the capeml ecosystem so is not meant to be called
directly (but can be).

Missing value codes are not documented for geometry columns of spatial
(sf) objects.
}
