% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_attributes_file.R
\name{read_attributes_file}
\alias{read_attributes_file}
\title{Load attribute information from a entity attribute yaml or csv file}
\usage{
read_attributes_file(string_pointer, entity_id)
}
\arguments{
\item{string_pointer}{(character) The quoted name of the data entity.}

\item{entity_id}{(character) Quoted identifier of the data object that is being described,
this will usually be the name or hash of the data table (or otherwise) of
which the attribute is associated.}
}
\value{
A entity of type EML attributes or list of attributes (for testing
and debugging)
}
\description{
The \code{read_attributes_file} function reads an entity's attribute
details from a "entity name"_attrs.yaml or "entity name"_attrs.csv file in
the working directory - the details of which are incorporated into the EML
metadata for that entity. If present in the working directory, factor
metadata in a "entity name"_factors.yaml or "entity name"_factors.csv are
read and incorporated into attribute metadata. NA, NaN, and user-provided
missing value codes are documented for each variable if they exist.
}
\details{
The \code{read_attributes_file} function reads an entity's attribute
details from a "entity name"_attrs.yaml or "entity name"_attrs.csv file in
the working directory - the details of which are incorporated into the EML
metadata for that entity. If present in the working directory, factor
metadata in a "entity name"_factors.yaml or "entity name"_factors.csv are
read and incorporated into attribute metadata. NA, NaN, and user-provided
missing value codes are documented for each variable if they exist.
}
\note{
Recent versions of the capeml package generate attribute and factor
metadata files in yaml format; the code{read_attributes_file} function will look also
for attributes files in csv format to accommodate older projects.

The \code{read_attributes_file} function is intended primarily as a helper to
other functions in the capeml ecosystem so is not meant to be called
directly (but can be).
}
