% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_raster_attributes.R
\name{read_raster_attributes}
\alias{read_raster_attributes}
\title{Read raster attribute information from attributes metadata file}
\usage{
read_raster_attributes(entity_name, entity_id = "data_entity")
}
\arguments{
\item{entity_name}{(character) The quoted name of the data entity.}

\item{entity_id}{(character) Quoted identifier of the data object that is being described,
this will usually be the name or hash of the data table (or otherwise) of
which the attribute is associated.}
}
\value{
A entity of type EML attributes or list of attributes (for testing
and debugging)
}
\description{
The \code{read_raster_attributes} function reads a raster's
attribute details from a "entity name"_attrs.yaml or "entity name"_attrs.csv
file in the working directory - the details of which are incorporated into
the EML metadata for that entity.

The \code{read_raster_attributes} function reads a raster's
attribute details from a "entity name"_attrs.yaml or "entity name"_attrs.csv
file in the working directory - the details of which are incorporated into
the EML metadata for that entity.
}
\note{
Recent versions of the capeml package generate attribute and factor
metadata files in yaml format; the \code{read_raster_attributes} function
will look also for attributes files in csv format to accommodate older
projects.

The \code{read_raster_attributes} function is intended primarily as a
helper to other functions in the capeml ecosystem so is not meant to be
called directly (but can be).
}
