% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_taxonomicCoverage.R
\name{set_taxonomic_coverage}
\alias{set_taxonomic_coverage}
\title{Create the taxonomicCoverage EML node}
\usage{
set_taxonomic_coverage(sci_names)
}
\arguments{
\item{sci_names}{(list) Object returned by \code{capeml::create_taxa_list()}.}
}
\value{
\item{list}{ a emld list object is returned for use with the EML R Package }
\item{.xml file}{ a .xml file is written to the working directory }
}
\description{
Create the taxonomicCoverage EML node
}
\note{
courtesy of the taxonomyCleanr package
}
\keyword{internal}
