% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_taxonomicCoverage.R
\name{set_taxonomicCoverage}
\alias{set_taxonomicCoverage}
\title{Generate EML taxonomic coverage element from a list of resolvable taxa}
\usage{
set_taxonomicCoverage(sci_names, expand = FALSE, db = "itis")
}
\arguments{
\item{sci_names}{A list of dataset taxa}

\item{expand}{Flag: set to TRUE to use taxize to expand sci_names into full
taxonomic classifications (FALSE is default)}

\item{db}{Information source for taxonomic details (ITIS is default)}
}
\value{
EML entity of type taxonomicCoverage
}
\description{
This is a stand-alone function borrowed from the rOpenSci EML
  package (https://github.com/ropensci/EML). Provided with a list of taxa
  names with resolvable ITIS identifiers, set_taxonomicCoverage generates a
  EML entity of type taxonomicCoverage for inclusion in an EML document.
}
\details{
Input is a list of ITIS-resolvable taxa (identify_resolvable_taxa
  should be run first and set_taxonomicCoverage run on that list). Output is
  a EML entity of type taxonomicCoverage for those taxa.
}
\examples{
\dontrun{

taxaCoverage <- set_taxonomicCoverage(resolved_taxa, expand = T, db = 'itis')

# set taxonomicCoverage element in EML coverage
# coverage@taxonomicCoverage <- c(taxaCoverage)

}

}
