% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_attributes.R
\name{update_attributes}
\alias{update_attributes}
\title{Update minimum and maximum values of numeric variables in existing
attribute metadata}
\usage{
update_attributes(entity_name, return_type = "yaml")
}
\arguments{
\item{entity_name}{(character) The name of the data entity.}

\item{return_type}{(character) Quoted designator indicating the value returned as either a
attributes template yaml file (return_type = "yaml", the default) or a list
of entity attributes (return_type = "attributes") constructed from the data
entity. The latter (i.e., return_type = "attributes") is really just for
testing the function.}
}
\value{
An updated metadata template including refreshed minimum and maximum
vales for numeric variables for providing attribute metadata as a yaml file
with the file name of the R data object + "_attrs.yaml" in the working
directory.
}
\description{
A frequent need with long-term, ongoing research is to update
existing data. A challenge to that is that we do not want to have to rebuild
from scratch the attribute metadata for a data entity at each update. In
terms of attribute metadata, definitions, units, etc. are relatively static
but what can change are the minimum and maximum values for numeric variables
as the observation record grows. We could ascertain the minimum and maximum
values for numeric variables then manually update existing attribute
metadata but this is tedious, error-prone, and can be time consuming when
dealing with many variables. The \code{update_attributes} function takes
care of this for us by reading the existing attribute metadata for a given
data entity and updating those metadata with the minimum and maximum values
for said data entity if they have changed in the context of a data refresh.
}
\note{
An artifact of the updating process is that the \code{definition}
element is populated in the updated yaml file. The \code{definition} element
is required by the EML schema for attributes of type character. In a typical
workflow, the \code{definition} element is left blank in the metadata yaml
file and is automatically populated with a copy of the
\code{attributeDefinition} element at build time (e.g.,
\code{create_dataTable}). Because \code{update_attributes} is calling
\code{read_attributes}, which is what populates any empty \code{definition}s
for variables of type character, this is reflected by
\code{update_attributes} in the updated yaml. This does not have any bearing
on the resulting data entity EML, and we can still provide custom
\code{definition} metadata if desired.

\code{update_attributes} will abort the update if an attribute is
detected in the data entity but for which there is not metadata in the
existing attributes file. This is an indication that the data structure or
content has changed sufficiently that a new, blank attributes metadata file
should be constructed. Conversely, if an attribute is detected in the
existing metadata that is not detected in the data entity, the update will
proceed but the attribute and corresponding metadata in the entity but not
in the existing metadata file will be stricken from the updated attribute
metadata file. In both cases, \code{update_attributes} will print to screen
the incongruent attributes.
}
\examples{
\dontrun{

 # update attributes file for mycars data object

 mycars <- head(mtcars)

 capeml::update_attributes(entity_name = mycars)

}

}
