% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_cap_eml.R
\name{write_cap_eml}
\alias{write_cap_eml}
\title{write eml to file}
\usage{
write_cap_eml(
      emlEntity,
      filePathName
    )
}
\arguments{
\item{emlEntity}{(unquoted character) Unquoted name of the eml entity (from
`create_eml`) to be written to file. Defaults to `eml`.}

\item{filePathName}{(character) Path and file name of file to be written. Defaults to the
package identifier and current directory.}
}
\value{
Writes eml object as specified name and to specified location in xml
    format.
}
\description{
A wrapper around EML::write_eml that automates file naming from projet
    configuration
}
\details{
Any existing xml files named with \code{packageScopeNumber} at
    \code{path} will be overwritten.
}
\examples{
\dontrun{

write_cap_eml(
  emlEntity = myeml,
  filePathName = "~/Desktop/myeml.xml"
)

write_cap_eml()

}

}
