% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_directory.R
\name{write_directory}
\alias{write_directory}
\title{construct campel supporting files in a data package directory}
\usage{
write_directory(
      scope,
      identifier,
      path
    )
}
\arguments{
\item{scope}{(character) Quoted name of the package scope (e.g., "edi"). The default is
"knb-lter-cap".}

\item{identifier}{(integer) Data package identifier (number).}

\item{path}{(character) Path to where the config file will be written. Defaults to the
current directory.}
}
\value{
A directory with the following structure and contents:
    \itemize{
        \item{\strong{name} Name supplied via \code{scope} and \code{identifier}}
        \itemize{
            \item{\strong{config.yaml} configuration file}
            \item{\code{data package name}.Rmd workflow template}
        }
    }
}
\description{
Create a directory structure for data package contents and
\code{capeml} files.
}
\details{
Existing directories named with \code{data package name} at
\code{path} will not be overwritten.
}
\examples{
\dontrun{
# Template data package directory "edi.521"

capeml::write_directory(
  scope      = "edi",
  identifier = 521,
  path       = '~/Desktop'
)

}

}
