% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_people_template.R
\name{write_people_template}
\alias{write_people_template}
\title{write a template yaml file for adding personnel (people) metadata}
\usage{
write_people_template(path = ".", overwrite = FALSE)
}
\arguments{
\item{path}{(character) Path to where the yaml file will be written. Defaults to the
current directory.}

\item{overwrite}{(logical) Indicates to overwrite an existing file if one exists.}
}
\value{
A people.yaml file in the prescribed directory for adding personnel
metadata; initially populated with inputs for a single <creator>,
<metadataProvider>, and <associatedParty>
}
\description{
Write a template yaml file (people.yaml) for adding pesonnel
(people) metadata
}
