% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_template.R
\name{write_template}
\alias{write_template}
\title{build a template workflow}
\usage{
write_template(
      path,
      packageScopeNumber,
      overwrite
    )
}
\arguments{
\item{packageScopeNumber}{(character) Quoted name of the package scope and number without the
version number (e.g., "edi.521").}

\item{path}{(character) Path to where the data package directory will be created.
Defaults to the current directory.}

\item{overwrite}{(logical) indicates to overwrite an existing file if one exists;
default = FALSE}
}
\value{
A Rmarkdown (Rmd) file named with the project scope and number
}
\description{
Write a workflow template Rmd file for the capeml workflow
}
\details{
Existing files named with \code{packageScopeNumber}.Rmd at \code{path}
    will not be overwritten unless overwrite is made explicit.
}
\examples{
\dontrun{
# Template data package directory "edi.521"

write_template(
  packageScopeNumber = "edi.521",
  path = '~/Desktop'
)

write_template(
  packageScopeNumber = "edi.521"
)

}

}
