% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_unit_dictionary.R
\name{write_unit_dictionary}
\alias{write_unit_dictionary}
\title{create EML standard unit dictionary}
\usage{
write_unit_dictionary()
}
\description{
A convenience function to easily generate the EML standard unit
    dictionary as a data frame.
}
