% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipRelatedFiles.R
\name{zipRelatedFiles}
\alias{zipRelatedFiles}
\title{zipRelatedFiles}
\usage{
zipRelatedFiles(targetFile)
}
\arguments{
\item{targetFile}{a spatial data file}
}
\value{
the zipped file name for inclusion in the EML documentation
}
\description{
the zipped file name for inclusion in the EML documentation is
  returned. Side effects include all relevant files zipped into a single,
  compressed file named as: projectid + basefile name + md5sum + extension
  (zip).
}
\details{
zipRelatedFiles is a helper function designed primarily to assist
  the creation of EML spatial data objects. A spatial data entity often
  consists of more than one file (e.g., shapefiles; raster (*.img, *.xml))
  that are incorporated into a single zipped file for upload to PASTA. This
  function zips related files as determined by the same base file name (e.g.,
  CAP_1985.img and CAP_1985.img.aux.xml share the same base name and would be
  zipped). In keeping with the CAP LTER workflow for generating EML metadata,
  the resulting compressed file is named projectid + basefile name + md5sum +
  extension (zip).
}
