% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_caplter.R
\docType{data}
\name{capRights}
\alias{capRights}
\title{capRights: default intellectual rights statement passed to EML-generating workflow}
\format{An object of class \code{character} of length 1.}
\usage{
capRights
}
\description{
capRights: default intellectual rights statement passed to EML-generating workflow
}
\keyword{datasets}
