% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_caplter.R
\docType{data}
\name{capRights}
\alias{capRights}
\title{CAP LTER default access / copy rights}
\format{An object of class \code{character} of length 1.}
\usage{
capRights
}
\description{
CAP LTER default access / copy rights
}
\keyword{datasets}
