% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dataset.R
\name{create_dataset}
\alias{create_dataset}
\title{create a EML entity of type dataset}
\usage{
create_dataset(publication_date = NULL)
}
\arguments{
\item{publication_date}{(character) Quoted ISO date - defaults to \code{Sys.Date()}}
}
\value{
EML entity of type dataset entity is returned.
}
\description{
\code{create_dataset} generates a EML entity of type dataset.
}
\details{
A dataset entity, the central component of a data package, is
created from objects in the user's R environment and supporting yaml files.
Prior to processing, any existing annotations.yaml and custom_units.yaml in
the working directory are removed silently to avoid stale metadata.
A project (default is LTER) indicates contact and project details specific
to the research. The abstract and methods must be in markdown format -
package will look for these files (abstract.md, methods.md) in the project
directory. Similarly for keywords, the package will look for a keywords.csv
file in the project directory.
}
\note{
\code{create_dataset} will look for most inputs used to construct a
dataset, such as a package number, in the working environment or from the
project config.yaml; these parameters are not passed directly to the
function and must exist in the working environment or configuration file(s).
}
