% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dataset.R
\name{create_dataset}
\alias{create_dataset}
\title{create_dataset}
\usage{
create_dataset(
  scope = "LTER",
  abstractFile = "abstract.md",
  methodsFile = "methods.md",
  keywordsFile = "keywords.csv",
  publicationDate = NULL
)
}
\arguments{
\item{scope}{Quoted name or acronym of the project.}

\item{abstractFile}{Quoted name and path of abstract (in markdown format)}

\item{methodsFile}{Quoted name and path of methods (in markdown format)}

\item{keywordsFile}{Quoted name and path of keywords (in csv format)}

\item{publicationDate}{Quoted ISO date - defaults to today's date}
}
\value{
EML dataset entity is returned.
}
\description{
create_dataset generates a EML entity of type dataset
}
\details{
A dataset entity, the central component of a data package, is
  created from objects in the user's R environment. A project scope (default
  is LTER) indicates contact and project details specific to the research.
  The abstract and methods must be in markdown format - by default the
  package will look for these files (abstract.md, methods.md) in the project
  directory but files of different names or locations can be passed.
  Similarly for keywords, the package will look for a keywords.csv file in
  the project directory but a different name or location can be passed.
}
\note{
create_dataset will look for most inputs used to construct a dataset,
  such as a project id, in the working environment; these parameters are not
  passed directly to the function and must exist in the working environment.
}
