% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_keywordSet.R
\name{create_keywordSet}
\alias{create_keywordSet}
\title{create_keywordSet}
\usage{
create_keywordSet(keywordsFile)
}
\arguments{
\item{keywordsFile}{The quoted name and path of the keywords file.}
}
\value{
EML keywordSet object is returned.
}
\description{
create_keywordSet generates a EML entity of type
  keywordSet.
}
\details{
A keywordSet entity is created from a single data file (.csv) with
  the fields: thesaurus, keyword, and type where type is an intended (but
  optional) attribute for the keyword.
}
\examples{
\dontrun{

datasetKeywordSet <- create_keywordSet('keywordTest.csv')

The workflow harvests keyword metadata from a keyword file (.csv) with the
columns: thesaurus, keyword, type where type is an optional attribute that
will be associated with the keyword EML/XML element.

}

}
