% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_package.R
\name{create_package}
\alias{create_package}
\title{create or update data package with EDI API}
\usage{
create_package(identifier = NULL, scope = NULL, environment = "staging")
}
\arguments{
\item{identifier}{(integer) Unquoted package identifier integer}

\item{scope}{(character) Quoted package scope (e.g., "knb-lter-cap", "edi")}

\item{environment}{(character) Quoted name of the EDI environment where the data package
should be created, can be: production, staging (default), or development.}
}
\description{
\code{create_package} creates a data package in the EDI data
repository.
}
\note{
\code{create_package} will look for a config.yaml file to estalish the
package scope and identifier. Alternatively, these parameters can be
provided by the user but, in that case, both parameters must be provided.

The user must have established a connection to the EDI API (see
EDIutils::login) in the current R session.
}
