% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_spatialRaster.R
\name{create_spatialRaster}
\alias{create_spatialRaster}
\title{create_spatialRaster}
\usage{
create_spatialRaster(rasterFile, description, epsgProjection,
  emlProjection, rasterValueDescription, rasterValueUnits,
  zipFiles = FALSE,
  baseURL = "https://data.gios.asu.edu/datasets/cap/")
}
\arguments{
\item{rasterFile}{Quoted full path to raster file.}

\item{description}{Description of the raster.}

\item{epsgProjection}{EPSG numeric code of raster's coordinate reference system}

\item{emlProjection}{EML-compliant refence to raster's coordinate reference system}

\item{rasterValueDescription}{Description of raster values}

\item{rasterValueUnits}{Raster value units}

\item{zipFiles}{Logical indicating whether spatial raster entity should be
constructed from a single raster file (FALSE, default) or entire directory (TRUE)}

\item{baseURL}{(optional) The base path of the web-accessible location of the
data file; the name of the resulting file will be passed to the base path
to generate a web-resolvable file path.}
}
\value{
EML spatial data object is returned. Additionally, the spatial data
  file is renamed with the project id + base file name + md5sum + file
  extension (zip in the case when multiple files are aggregated).
}
\description{
create_spatialRaster generates a EML entity of type
  spatialRaster
}
\details{
a spatialRaster entity is created from a single data file (e.g.,
  CAP_1985.img) or a collection of related files (e.g., CAP_1985.img,
  CAP_1985.img.aux.xml). In the case of multiple files, all files in the
  parent directory where the raster file is located are aggregated into a
  single compressed (zipped) file. In all cases, the resulting entity is
  renamed with the project id + base file name + md5sum + file extension (zip
  in the case when multiple files are aggregated).
}
\note{
create_spatialRaster will look for a project id in the working
  environment; this parameter is not passed to the function and it must
  exist.
}
\examples{
\dontrun{
The workflow features harvesting metadata from template files. Most metadata
are documented in a raster-level metadata file (e.g., CAP1985_metadata.csv
below). If raster values are categorical, an addiitonal metadata file is
required to document the categories (e.g., CAP1985_factors.csv below)

 spatial_entity <- create_spatialRaster(
   pathToRaster = "~/folder/",
   metadataFile = "~/folder/CAP1985_metadata.csv",
   categoricalMetadataFile = "~/folder/CAP1985_factors.csv")
}

}
