% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_spatialRaster.R
\name{create_spatialRaster}
\alias{create_spatialRaster}
\title{create_spatialRaster}
\usage{
create_spatialRaster(pathToRaster, metadataFile, categoricalMetadataFile)
}
\arguments{
\item{pathToRaster}{The quoted path of the raster directory.}

\item{metadataFile}{The quoted path and name of the raster metadata file (the
name of the raster data to be processed is included as a parameter in this
metadata file).}

\item{categoricalMetadataFile}{Included if raster values are categorical. The
quoted path and name of the raster metadata file that provides said
details.}
}
\value{
EML spatial data object is returned. Additionally, the spatial data
  file is renamed with the project id + base file name + md5sum + file
  extension (zip in the case when multiple files are aggregated).
}
\description{
create_spatialRaster generates a EML entity of type
  spatialRaster
}
\details{
a spatialRaster entity is created from a single data file (e.g.,
  CAP_1985.img) or a collection of related files (e.g., CAP_1985.img,
  CAP_1985.img.aux.xml). In the case of multiple file, all related files as
  identified by a common base name (e.g., 'CAP_1985') are aggregated into a
  single compressed (zipped) file. In all cases, the resulting entity is
  renamed with the project id + base file name + md5sum + file extension (zip
  in the case when multiple files are aggregated).
}
\note{
create_spatialRaster will look for a project id in the working
  environment; this parameter is not passed to the function and it must
  exist.

create_spatialRaster currently does not accept an argument for url to a
  data file; the package defaults to a CAP LTER specific URL, and this must
  be changed manually in the resulting xml

create_spatialRaster relies on the helper functions zipRelatedFiles and
  get_emlProjection, which are specific to UNIX operating systems.
}
\examples{
\dontrun{
The workflow features harvesting metadata from template files. Most metadata
are documented in a raster-level metadata file (e.g., CAP1985_metadata.csv
below). If raster values are categorical, an addiitonal metadata file is
required to document the categories (e.g., CAP1985_factors.csv below)

 spatial_entity <- create_spatialRaster(
   pathToRaster = "~/folder/",
   metadataFile = "~/folder/CAP1985_metadata.csv",
   categoricalMetadataFile = "~/folder/CAP1985_factors.csv")
}

}
