% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_caplter.R
\docType{data}
\name{giosAddress}
\alias{giosAddress}
\title{configure_caplter}
\format{An object of class \code{list} of length 8.}
\usage{
giosAddress
}
\value{
establishes in the R ennvironment upon package load, eml entities
  referenced in (or outside) the remlTemplate.
}
\description{
configure_caplter establishes several data set elements that are
  likely to be consistent across all CAP LTER data sets, including the GIOS
  address, contact information, landuage, intellectual rights, access, and
  metadata distribution.
}
\details{
configure_caplter establishes several data set elements that are
  likely to be consistent across all CAP LTER data sets, including the GIOS
  address, contact information, landuage, intellectual rights, access, and
  metadata distribution. Most of these eml entities are established in the R
  environment when the capeml package is loaded and available for references
  in the remlTemplate. Distribution is the exception as this is a function
  that requires the data set package identifier as an input.
}
\examples{
\dontrun{

# most entites loaded when the package is loaded, except distrubution.

create_distribution(packageIdent)

}

}
\keyword{datasets}
