% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_caplter.R
\docType{data}
\name{giosLanguage}
\alias{giosLanguage}
\title{giosLanguage: default language (english) passed to EML-generating workflow}
\format{
An object of class \code{character} of length 1.
}
\usage{
giosLanguage
}
\description{
giosLanguage: default language (english) passed to EML-generating workflow
}
\keyword{datasets}
