% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_caplter.R
\docType{data}
\name{giosPublisher}
\alias{giosPublisher}
\title{default CAP LTER / GIOS publisher as EML::publisher}
\format{An object of class \code{list} of length 11.}
\usage{
giosPublisher
}
\description{
default CAP LTER / GIOS publisher as EML::publisher
}
\keyword{datasets}
