% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_resolvable_taxa.R
\name{identify_resolvable_taxa}
\alias{identify_resolvable_taxa}
\title{Identify taxa within a list that can be matched to ITIS}
\usage{
identify_resolvable_taxa(taxa_list)
}
\arguments{
\item{taxa_list}{A list of taxa to validate}
}
\value{
List of provided taxa and associated ITIS identifier if a match was
  found (NA if otherwise).
}
\description{
Provided with a list of taxa names, identify_resolvable_taxa
  generates a list of those taxa and the associated ITIS identifier if a
  match was identified. identify_resolvable_taxa is intended primarily as a
  helpful function for set_taxonomicCoverage but can be used independently.
}
\details{
A list of provided taxa and associated ITIS identifier if a match
  could be found.
}
\examples{
\dontrun{

resolvedTaxa <- create_keyword(c('Ephemeroptera', 'Plecoptera')
resolvedTaxa <- create_keyword(listOfTaxa)

}

}
