% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_caplter_project.R,
%   R/configure_urex_project.R
\docType{data}
\name{projectTitle}
\alias{projectTitle}
\title{configure_caplter_project}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
projectTitle

projectTitle
}
\value{
establishes in the R environment upon package load a EML::project
  entity specific to the CAP LTER. If appropriate to include in a dataset,
  'capProject' should be referenced with the project argument when
  constructing a EML::dataset.

establishes in the R environment upon package load a EML::project
  entity specific to the UREx project. If appropriate to include in a
  dataset, 'urexProject' should be referenced with the project argument when
  constructing a EML::dataset.
}
\description{
configure_caplter_project establishes an EML::project with the
  elements: (1) personnel (CAP PI and Co-PIs), (2) project abstract (project
  overview from CAP IV proposal), and (3) funding (grant numbers for CAP I,
  II, III, and IV).

configure_urex_project establishes an EML::project with the
  elements: (1) personnel, (2) project abstract, and (3) funding
}
\details{
configure_caplter_project establishes an EML::project with the
  elements: (1) personnel (CAP PI and Co-PIs), (2) project abstract (project
  overview from CAP IV proposal), and (3) funding (grant numbers for CAP I,
  II, III, and IV).

configure_urex_project establishes an EML::project with the
  elements: (1) personnel, (2) project abstract, and (3) funding
}
\keyword{datasets}
