% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_attributes.R
\name{write_attributes}
\alias{write_attributes}
\title{write_attributes}
\usage{
write_attributes(dfname, MVCE, overwrite = "no")
}
\arguments{
\item{dfname}{The unquoted name of the R data frame or Tibble.}

\item{MVCE}{an optional missing value code explanation (MVCE). If MVCE is
omitted, default text of `missing value` is used to document missing values
(always NAs).}

\item{overwrite}{Quoted yes or no indicating if an existing attributes file
in the target directory should be overwritten.}
}
\value{
The name of the file generated is returned, and a template for
  providing attribute metadata as a csv file with the file name of the R data
  object + "_attrs.csv" is created in the working directory.
}
\description{
write_attributes creates a template as a csv file for supplying
  attribute metadata for a tabular data object that resides in the R
  environment.
}
\details{
The csv template generated by write_attributes includes the field
  names of the data entity. The number type, column class (e.g., factor,
  numeric), minimum and maximum values (if numeric), and missing value code
  and explanation (if provided) for each field. The template supports input
  of format string, unit, definition, and attribute definition. The csv file
  is written with the name of the data object in R + "_attrs". The
  create_dataTable function will search for this file will creating a EML
  dataTable entity.
}
\examples{
\dontrun{

 write_attributes(R data object)
 write_attributes(R data object, overwrite = 'yes')

}

}
