% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_config.R
\name{write_config}
\alias{write_config}
\title{write a project config.yaml file}
\usage{
write_config(scope = "knb-lter-cap", identifier, path = ".", overwrite = FALSE)
}
\arguments{
\item{scope}{(character) Quoted name of the package scope (e.g., "edi"). The default is
"knb-lter-cap".}

\item{identifier}{(integer) Data package identifier (number).}

\item{path}{(character) Path to where the config file will be written. Defaults to the
current directory.}

\item{overwrite}{(logical) Logical indicating if an existing config file in the target
directory should be overwritten.}
}
\description{
write_config writes a template config.yaml file to the working
directory (default) or supplied path.
}
\details{
A template config.yaml file is written to the working directory or
specified location. The function requires that the package scope and number
(e.g., "edi", 521) are provided. Package identifiers include a verion
number; the function uses a default value of version 1 but an alternate
version number can be specified.
}
\note{
The function expects specifically a 3-digit package identifier (number).
}
