% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_template.R
\name{write_template}
\alias{write_template}
\title{Construct a template capeml workflow}
\usage{
write_template(
      scope,
      identifier,
      path,
      overwrite
    )
}
\arguments{
\item{scope}{(character) Quoted name of the package scope (e.g., "edi"). The default is
"knb-lter-cap".}

\item{identifier}{(integer) Data package identifier (number).}

\item{path}{(character) Path to where the config file will be written. Defaults to the
current directory.}

\item{overwrite}{(logical) indicates to overwrite an existing file if one exists (default =
FALSE).}
}
\value{
A Rmarkdown (Rmd) file named with the project scope and identifier
(number).
}
\description{
Write a workflow template Rmd file for the capeml workflow
}
\details{
Existing files with the package name and .Rmd extension at
\code{path} will not be overwritten unless overwrite is made explicit.
}
\examples{
\dontrun{
# Template data package directory "edi.521"

write_template(
  scope      = "edi",
  identifier = 521,
  path       = '~/Desktop'
)

}

}
