% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_template.R
\name{write_template}
\alias{write_template}
\title{Construct a template capeml workflow}
\usage{
capeml::write_template(
      scope,
      identifier,
      path,
      overwrite
    )
}
\arguments{
\item{scope}{(character) Quoted name of the package scope (e.g., "edi"). The default is
"knb-lter-cap".}

\item{identifier}{(integer) Data package identifier (number).}

\item{path}{(character) Path to where the config file will be written. Defaults to the
current directory.}

\item{overwrite}{(logical) indicates to overwrite an existing file if one exists (default =
FALSE).}
}
\value{
A quarto (qmd) file named with the project scope and identifier
(number).
}
\description{
Write a workflow template qmd file for the capeml workflow
}
\details{
Existing files with the package name and .qmd extension at
\code{path} will not be overwritten unless overwrite is made explicit.
}
\note{
While the function does provide a generalized template for much of the
capeml workflow, creating data entities (e.g., dataTable, otherEntity) are
not included. Further, the template is geared heavily to drawing on a
metadata [template](https://globalfutures.asu.edu/caplter/data/submit-data/)
that is part of the data submission process for the CAP LTER.
}
\examples{
\dontrun{
# Template data package directory "edi.521"

capeml::write_template(
  scope      = "edi",
  identifier = 521,
  path       = '~/Desktop'
)

}

}
