% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_unit_dictionary.R
\name{write_unit_dictionary}
\alias{write_unit_dictionary}
\title{EML standard unit dictionary data frame}
\usage{
write_unit_dictionary()
}
\description{
A convenience function to easily generate the EML standard unit
    dictionary as a data frame.
}
